rename airQuality to air.py
