+++ src/pyfl/__init__.py
@@ -4,5 +4,5 @@ from .helper import *
 from .tube import *
 from .client import *
 from .accident import *
-from .airQuality import *
+from .air import *
 from .bikePoint import *
